+++ CHANGELOG.md
@@ -30,7 +30,9 @@
 - [BREAKING] Remove default value account filepath for faucet commands and rename --output-path to --output (#976).
 - Exposed `miden-proving-service` as a library (#956).
 - [BREAKING] Enforce `PoW` on all faucet API key-authenticated requests (#974).
+- Remove faucet rate limiter by IP and API Key, this has been superseded by PoW (#1011).
 - Faucet challenge expiration time is now configurable (#1017).
+- Faucet `PoW` difficulty is now configurable (#924).
 - Removed system monitor from node binary (#1019).
 - [BREAKING] Renamed `open_telemetry` to `enable_otel` in all node's commands (#1019).
 - [BREAKING] Rename `miden-proving-service` to `miden-remote-prover` (#1004).